api-gtwy-pipeline/terraform/modules/lambda/requirement.txt
botocore.vendored
freecurrencyapi
os
pprint
json
urllib3=1.26.15
